Fix giveFeedback ABI types (int128/uint8) and fragment basename collision
- Use int128 for value, uint8 for valueDecimals to match live ERC-8004
  registry selector
- Content-address served filenames (stem-hashprefix.ext) so publishing
  two files with the same basename doesn't overwrite the first

+++ contracts/src/interfaces/IReputationRegistry.sol
@@ -5,8 +5,8 @@ pragma solidity ^0.8.24;
 interface IReputationRegistry {
     function giveFeedback(
         uint256 agentId,
-        uint256 value,
-        uint256 valueDecimals,
+        int128 value,
+        uint8 valueDecimals,
         string calldata tag1,
         string calldata tag2,
         string calldata endpoint,
